Added TestLogger and TestLogger<T> stubs
